Move esmvaltool.interface_scripts.regrid to esmvaltool.preprocessor.regrid
