--- tests/integration/interface_scripts/regrid/__init__.py
@@ -1,4 +1,0 @@
-"""
-Integration tests for the :mod:`esmvaltool.interface_scripts.regrid` module.
-
-"""
