+++ tests/integration/preprocessor/regrid/__init__.py
@@ -1,0 +1,4 @@
+"""
+Integration tests for the :mod:`esmvaltool.preprocessor.regrid` module.
+
+"""
